+++ src/Type/FloatRangeT.php
@@ -22,7 +22,16 @@ final readonly class FloatRangeT implements Type
     public function __construct(
         public ?BigDecimal $min = null,
         public ?BigDecimal $max = null,
-    ) {}
+    ) {
+        if ($min !== null && $max !== null && $min->isGreaterThan($max)) {
+            throw new \ValueError(\sprintf(
+                '`%s` requires min to be less than or equal to max, got min=%d, max=%d',
+                self::class,
+                $min,
+                $max,
+            ));
+        }
+    }
 
     #[\Override]
     public function accept(Visitor $visitor): mixed