+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **BC Break:** Move `Typhoon\Type\Type` interface to `Typhoon\Type`.
 - **BC Break:** Require `Typhoon\Type\UnionT` to have at least 2 types.
 - **BC Break:** Require `Typhoon\Type\IntersectionT` to have at least 2 types.
+- **BC Break:** Require `$min <= $max` in `Typhoon\Type\IntRangeT`.
+- **BC Break:** Require `$min <= $max` in `Typhoon\Type\FloatRangeT`.
 
 ## [0.5.1]